+++ mlir/lib/CMakeLists.txt
@@ -6,7 +6,7 @@ add_flag_if_supported("-Werror=global-constructors" WERROR_GLOBAL_CONSTRUCTOR)
 add_subdirectory(Conversion)
 add_subdirectory(Dialect)
 #add_subdirectory(EDSC)
-add_subdirectory(ExecutionEngine)
+#add_subdirectory(ExecutionEngine)
 #add_subdirectory(IR)
 #add_subdirectory(Interfaces)
 #add_subdirectory(Parser)
