--- mlir/lib/ExecutionEngine/CMakeLists.txt
@@ -1,2 +1,0 @@
-# Placeholder to keep the original directory structure.
-add_subdirectory(ROCm)
